Recreate TabletsMetadata iterator when file ranges are not contiguous
In the Bulk Import v2 LoadFiles step a single TabletsMetadata
object was used to map a tables tablets to a set of bulk import
files. In the case where a small percentage of tablets were
involved in the bulk import a majority of the tables tablets
would still be evaluated. In the case where bulk imports were
not importing into contiguous tablets the code would just
iterate over the tables tablets until it found the next starting
point.

This change recreates the TabletMetadata object when a set of
files is not going to start at the next tablet in the table. A
likely better way to achieve the same thing would be to reset
the range on the underlying Scanner and create a new iterator,
but the TabletsMetadata object does not expose the Scanner. This
change also closes the TabletsMetadata objects which was not
being done previously.

Related to #5201